requirements.txt
pyglet>=1.3.2
pygame
PyOpenGL
numpy
noise
